+++ app/(main)/layout.tsx
@@ -3,8 +3,8 @@ import Link from "next/link"
 import { marketingConfig } from "@/config/marketing"
 import { cn } from "@/lib/utils"
 import { buttonVariants } from "@/components/ui/button"
+import { MainNav } from "@/components/main-nav"
 
-// import { MainNav } from "@/components/main-nav"
 // import { SiteFooter } from "@/components/site-footer"
 
 interface MarketingLayoutProps {
@@ -16,20 +16,9 @@ export default async function MarketingLayout({
 }: MarketingLayoutProps) {
   return (
     <div className="flex min-h-screen flex-col">
-      <header className="container z-40 bg-background">
-        <div className="flex h-20 items-center justify-between py-6">
-          {/* <MainNav items={marketingConfig.mainNav} /> */}
-          <nav>
-            <Link
-              href="/login"
-              className={cn(
-                buttonVariants({ variant: "secondary", size: "sm" }),
-                "px-4"
-              )}
-            >
-              Login
-            </Link>
-          </nav>
+      <header className="px-8 z-40 bg-background">
+        <div className="flex h-20 items-center py-6">
+          <MainNav items={marketingConfig.mainNav} />
         </div>
       </header>
       <main className="flex-1">{children}</main>
