+++ hooks/use-mounted.ts
@@ -1,0 +1,11 @@
+import * as React from "react"
+
+export function useMounted() {
+  const [mounted, setMounted] = React.useState(false)
+
+  React.useEffect(() => {
+    setMounted(true)
+  }, [])
+
+  return mounted
+}
