+++ types/index.d.ts
@@ -1,3 +1,11 @@
+export type NavItem = {
+  title: string
+  href: string
+  disabled?: boolean
+}
+
+export type MainNavItem = NavItem
+
 export type SiteConfig = {
   name: string
   description: string
